fix(platform): smart truncate gitlab pr body
